output: XML: add infix in output file
This patch adds an infix to the XML file to avoid problems if we are
logging packets and flows at the same time. Thus, we create two
different XML files whose filename describes the sort of logging
information that it contains. It is also useful when listing files
at a quick sight.

Signed-off-by: Pablo Neira Ayuso <pablo@netfilter.org>